% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoClassSurv.R
\name{oncoClassSurv}
\alias{oncoClassSurv}
\title{oncoClassSurv}
\usage{
oncoClassSurv(
  exp.type = "fpkm",
  train.exp.path = system.file("extdata", paste0("train.tumor.exp.", exp.type, ".txt"),
    package = "oncoClassSurv"),
  train.clin.path = system.file("extdata", "train.cluster.surv.rds", package =
    "oncoClassSurv"),
  train_cluster.feature.path = system.file("extdata", "train_cluster.features.rds",
    package = "oncoClassSurv"),
  train_survival.feature.path = NULL,
  input.exp.path,
  input.clin.path = NULL,
  miss_go.on = TRUE,
  rm.batch.effect = TRUE,
  task = 3,
  plot.combatch = FALSE,
  print.combat.plots = FALSE,
  cluster.method = "RF",
  nodesize = 3,
  ntree = 3000,
  mtry = 5,
  importance = T,
  kernel = "radial",
  cost = 4,
  surv.t.custom = NULL,
  time = "OS",
  event = 1,
  plot.surv.curve = TRUE,
  plot.samples = c(1:5),
  survtime.unit = "months",
  survcurve.break.x.by = 12,
  print.survplot = FALSE
)
}
\arguments{
\item{exp.type}{A character for the normalized format of RNA-Seq expression matrix. The default
is "fpkm", and the other option is "tpm". Please use the consistent
"exp.type" in both train expression matrix and input expression matrix.}

\item{train.exp.path}{The path of the RNA-Seq expression profile of the training cohort.
The default path is 'system.file("extdata", paste0("train.tumor.exp.", exp.type,
".txt"),package = "oncoClassSurv")' depending on the "exp.type".
This path can be any other customized path as per the user's demands.}

\item{train.clin.path}{The path of a table file containing clinical information. When performing the
classification task, the information on clusters in the training cohort is
essential for machine learning. When performing the prediction for survival
risk, survival data of the training cohort is essential for fitting the Cox
regression model.}

\item{train_cluster.feature.path}{This option can be customized. The path of a table file containing the marker
genes of clusters. When performing the classification task, the cluster-
specific marker genes is essential for precise training. Theoretically,
users can use all genes in the expression profile, but this is not recommended.}

\item{train_survival.feature.path}{Optional item. The path of a table file (.txt, csv, xlsx, xls, etc. or .rds)
including all covariables for training in the Cox regression model, when
prediction for survival risk needs to be performed. The default value is null.}

\item{input.exp.path}{The path of the RNA-Seq expression profile of the user's cohort. Please use
the consistent normalized format per the expression matrix in the training
cohort.}

\item{input.clin.path}{Optional item. The path of a table file (.txt, csv, xlsx, xls, etc. or .rds)
including additional clinical variables (e.g. tumor stage, age, gender, etc.),
when prediction for the survival risk of input data needs to be performed.
The default value is null.}

\item{miss_go.on}{Logical value. The function will automatically check whether missing genes
exist in the user's input expression matrix. If there are missing genes,
selecting "TRUE" means using the common genes (the training expression
matrix and the user's input expression matrix) and continuing to perform
the current task. This may cause imprecise prediction.}

\item{rm.batch.effect}{Logical value. Whether need to remove the batch effects between
the expression matrix of the training cohort and the expression matrix input by the user.}

\item{task}{Number value. Perform prediction for classifications when the task is 1.
Perform prediction for survival risk when the task is 2. Perform prediction for classifications and survival risk when the task is 3. The default value is 3.}

\item{plot.combatch}{Logical value. After removing the batch effects, visualize and return the
comparison results. Using the principal component analysis. The default value is FALSE.}

\item{print.combat.plots}{Logical value. After removing the batch effects, visualize the comparison
results in the current graph panel. This item works when the item of "plot.combatch" is TRUE.}

\item{cluster.method}{A character indicating the algorithm used in the prediction for classification.
Optional values include "RF" (random forest) and "SVM" (support vector machine).}

\item{nodesize}{A parameter in the random forest. Details can be seen in
the randomForest::randomForest().}

\item{ntree}{A parameter in the random forest. Details can be seen in the
randomForest::randomForest().}

\item{mtry}{A parameter in the random forest. Details can be seen in the
randomForest::randomForest().}

\item{importance}{A parameter in the random forest. Details can be seen in the
randomForest::randomForest().}

\item{kernel}{A parameter in the support vector machine. Details can be seen in the
e1071::svm().}

\item{cost}{A parameter in the support vector machine. Details can be seen in the
e1071::svm().}

\item{surv.t.custom}{When predicting the survival risk of patients, users can specify
concerned time points. The default value is null, which means predicting for
all time points.}

\item{time}{A character indicating the categories of survival time, such as overall
survival (OS), recurrence-free survival (RFS), etc. The default value is "OS".}

\item{event}{A character or number indicating the categories of survival event, such as
death or recurrence customized by users. The default value is 1.}

\item{plot.surv.curve}{A logical value. Whether return curves of survival risk over time
for individual patients. The default value is TRUE.}

\item{plot.samples}{A number. When there are too many patients, plotting multiple survival curves
on one graph panel will appear too crowded. Therefore, it is recommended to
select only the patients concerned by the user (default is numbered 1, 2, 3,
4, 5) to show their survival curves.}

\item{survtime.unit}{A character. The unit of the survival time, such as "days", "months", "year",
etc. as per the user's demands.}

\item{survcurve.break.x.by}{A number. The breakpoints of the time axis for the survival curves.
The default value is 12. Details can be seen in the survminer::ggsurvplot().}

\item{print.survplot}{A logical value. Whether visualize curves of survival risk over time for
individual patients in the current graph panel. This item works when the item
of "plot.surv.curve" is TRUE. The default value is FALSE.}
}
\value{
A list including classifications or survival risk results or both of them. The returned results depend on the task performed.
}
\description{
Classify tumors into several molecular sub-types using "random forest" or
"support vector machine" according to RNA-Seq data.
Visualization for survival risk based on Cox regression.
}
\examples{
\dontrun{
#perform task 1 (classify by SVM)
results<-oncoClassSurv(input.exp.path = system.file("extdata",
"icgc.tumor.exp.fpkm.txt",package = "oncoClassSurv"),
input.clin.path = system.file("extdata", "input_clinsurv.txt",
package = "oncoClassSurv"),miss_go.on=T,task=1,
rm.batch.effect=TRUE,plot.combatch=TRUE,
print.combat.plots=TRUE,cluster.method="SVM")
}

\dontrun{
#perform task 1 (classify by RF)
results<-oncoClassSurv(input.exp.path = system.file("extdata",
"icgc.tumor.exp.fpkm.txt",package = "oncoClassSurv"),
input.clin.path = system.file("extdata", "input_clinsurv.txt",
package = "oncoClassSurv"),task=1,rm.batch.effect=TRUE,
plot.combatch=TRUE,print.combat.plots=TRUE,cluster.method="RF")
}

\dontrun{
#perform task 2
results<-oncoClassSurv(train_survival.feature.path=system.file("extdata",
 "train_survival.features.rds",package = "oncoClassSurv"),
 input.exp.path = system.file("extdata", "icgc.tumor.exp.fpkm.txt",
 package = "oncoClassSurv"),input.clin.path = system.file("extdata",
 "input_clinsurv.txt",package = "oncoClassSurv"),
 task=2,rm.batch.effect=TRUE,plot.combatch=TRUE,
 print.combat.plots=TRUE,surv.t.custom=NULL,
 plot.surv.curve=TRUE,survcurve.break.x.by = 12,
 print.survplot = TRUE,plot.samples=c(1:5))
}

\dontrun{
#perform task 3
results<-oncoClassSurv(train_survival.feature.path=system.file("extdata",
"train_survival.features.rds",package = "oncoClassSurv"),
input.exp.path = system.file("extdata", "icgc.tumor.exp.fpkm.txt",
package = "oncoClassSurv"),input.clin.path = system.file("extdata",
"input_clinsurv.txt",package = "oncoClassSurv"),task=3,
rm.batch.effect=TRUE,plot.combatch=TRUE,print.combat.plots=TRUE,
cluster.method="SVM",surv.t.custom=NULL,plot.surv.curve=TRUE,
survcurve.break.x.by = 12,print.survplot = TRUE,plot.samples=c(1:5))
}


}
\author{
Yang Li
}
