% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncoClassSurv_RunShiny.R
\name{oncoClassSurv_RunShiny}
\alias{oncoClassSurv_RunShiny}
\title{oncoClassSurv_RunShiny}
\usage{
oncoClassSurv_RunShiny()
}
\value{
Classifications summary, heat map plot and/or tables from prediction results.
}
\description{
Perform convenient interactive operation through shiny application.
}
\examples{
\dontrun{
#Perform an interactive operation:
oncoClassSurv_RunShiny()
}
}
